Add test for addComponents()
